% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Image Renaming Application}
\usage{
run_app()
}
\value{
A Shiny app object
}
\description{
Launch the Shiny app for renaming and compressing nutrient deficiency images
}
\examples{
\dontrun{
run_app()
}
}
